requirements.txt
pandas==2.2.3
scikit-learn==1.5.2
psycopg2==2.9.9
SQLAlchemy==2.0.35
python-dotenv==1.0.1
joblib==1.4.2
